+++ docs/Changelog.md
@@ -35,6 +35,7 @@
 - Add `maxAreaNodes` configuration parameter for changing an area visibility calculation limit [#3501](https://github.com/opentripplanner/OpenTripPlanner/issues/3534)
 - Add bicycle safety report to report API [#3563](https://github.com/opentripplanner/OpenTripPlanner/issues/3563)
 - Optimize Transfers performance issue [#3513](https://github.com/opentripplanner/OpenTripPlanner/issues/3513)
+- Refactor StopPattern/TripPattern/TripTimes [#3571](https://github.com/opentripplanner/OpenTripPlanner/issues/3571)
 - Don't allow bicycle loops in A* [#3574](https://github.com/opentripplanner/OpenTripPlanner/pull/3574)